requirements.txt
requests==2.31.0
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
schedule==1.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.6